### 2022-10-27 #112 @whedon
[PRE REVIEW]: GraphPPL.jl: A Julia Package for probabilistic model and inference constraints specification
**Submitting author:** @bvdmitri (<a href="http://orcid.org/0000-0001-9655-7986">Dmitry Bagaev</a>)
**Repository:** <a href="https://github.com/biaslab/GraphPPL.jl" target ="_blank">https://github.com/biaslab/GraphPPL.jl</a>
**Version:**
