### 2022-11-16 c25c7e0
update write_randomprocess_options
