### 2022-11-24 d8325b5
feat: support for priors as arguments
