+++ versions.js
@@ -1,10 +1,11 @@
 var DOC_VERSIONS = [
   "stable",
+  "v3.1",
   "v3.0",
   "v2.3",
   "v2.2",
   "v2.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v3.0.0";
+var DOCUMENTER_NEWEST = "v3.1.0";
 var DOCUMENTER_STABLE = "stable";
