+++ v3.1.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"transformation-steps/#Model-specification-transformation-steps-for-the-ReactiveMP.jl-backend","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"","category":"section"},{"location":"transformation-steps/#Step-1:-Normalizarion-of-operator-node-arguments","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Step 1: Normalizarion of ~ operator node arguments","text":"","category":"section"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"Any expression of the form ","category":"page"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"lhs ~ Node(..., f(...), ...)","category":"page"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"is translated to","category":"page"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"lhs ~ Node(..., var\"#anonymous\" ~ f(...), ...)","category":"page"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"The only one exception is reference expression of the form x[f(i)] which are left untouched. This step forces model to create an anonymous node for any inner function call within ~ operator expression. In some cases ReactiveMP.jl backend can (and will) optimize this inner anonymous nodes into just function calls. E.g. following example won't create any additional nodes in the model ","category":"page"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"precision = 1.0\n\nnoise ~ NormalMeanVariance(noise_mean, 1.0 / precision) # Since 1.0 and precision are constants inference backend can just apply `/` function to them `/(1.0, precision)`.","category":"page"},{"location":"transformation-steps/#Step-2:-Main-pass","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Step 2: Main pass","text":"","category":"section"},{"location":"transformation-steps/#datavar()-transformation","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"datavar() transformation","text":"","category":"section"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"Any expression of the form ","category":"page"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"y = datavar(args...) # empty options here\n# or \ny = datavar(args...) where { options... }","category":"page"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the ReactiveMP.jl backend","text":"is translated to ","category":"page"},{"location":"transformation-steps/","page":"Model specification transformation steps for the ReactiveMP.jl backend","title":"Model specification transformation steps for the
[…]
