+++ src/model.jl
@@ -328,14 +328,14 @@ function generate_model_expression(backend, model_specification)
         elseif @capture(expression, varexpr_ = randomvar(arguments__) where { options__ })
             return :($varexpr = randomvar($(arguments...); $(options...)))
         elseif @capture(expression, varexpr_ = randomprocess(arguments__) where { options__ })
-            return :($varexpr = randomprocess($(arguments...); $(options...)))
+            return :($varexpr = randomprocess($(arguments...); $(options...)))   #add randomprocess here 
         elseif @capture(expression, varexpr_ = datavar(arguments__) where { options__ })
             return :($varexpr = datavar($(arguments...); $(options...)))
         elseif @capture(expression, varexpr_ = constvar(arguments__) where { options__ })
             return error("Error in the expression $(expression). `constvar()` call does not support `where {  }` syntax.")
         elseif @capture(expression, varexpr_ = randomvar(arguments__))
             return :($varexpr = randomvar($(arguments...); ))
-        elseif @capture(expression, varexpr_ = randomprocess(arguments__))
+        elseif @capture(expression, varexpr_ = randomprocess(arguments__))  #add randomprocess here 
             return :($varexpr = randomprocess($(arguments...); ))
         elseif @capture(expression, varexpr_ = datavar(arguments__))
             return :($varexpr = datavar($(arguments...); ))
@@ -377,8 +377,8 @@ function generate_model_expression(backend, model_specification)
             return write_randomvar_expression(backend, model, varexpr, rvoptions, arguments)
 
         elseif @capture(expression, varexpr_ = randomprocess(arguments__; options__))
-            rvoptions = write_randomprocess_options(backend, varexpr, options)
-            return write_randomprocess_expression(backend, model, varexpr, rvoptions, arguments)
+            rpoptions = write_randomprocess_options(backend, varexpr, options)
+            return write_randomprocess_expression(backend, model, varexpr, rpoptions, arguments)
         # Step 3.3 Convert constvar calls 
         elseif @capture(expression, varexpr_ = constvar(arguments__))
             return write_constvar_expression(backend, model, varexpr, arguments)
