### 2022-12-13 #112 @bvdmitri
The paper is in the [`paper/julia_proceedings_2022`](https://github.com/biaslab/GraphPPL.jl/tree/paper/julia_proceedings_2022) branch.
