### 2023-01-17 #112 @bvdmitri
@matbesancon  No, last commit has been made more than 4 months ago. I checked it now locally using the following instructions https://juliacon.github.io/proceedings-guide/author/#local_build and it works perfectly.
