### 2023-01-17 #112 @bvdmitri
@matbesancon Here is the pdf document generated locally
[paper.pdf](https://github.com/JuliaCon/proceedings-review/files/10437395/paper.pdf)
