### 2023-01-18 #112 @bvdmitri
@matbesancon  The official guideline mentions the [Whedon bot](https://whedon.theoj.org/), but it seems to me that it does not support `paper.tex` submissions, which is the only possible way for the JuliaCon Proceedings. Thus I have no possibility to make sure that the `whedon` works without errors. `whedon` docker image and GitHub Action do work with `paper.md` submissions, but I don't see any documentation on how to run them for `paper.tex` submissions.

The log posted in the previous message is uninformative and does not really help to identify the issue. Is there a way to get an extensive log?
