### 2023-01-23 #112 @matbesancon
@bvdmitri one thing to watch out: we need the `.latexrc` file, check the files here:
https://github.com/JuliaCon/JuliaConSubmission.jl/tree/master/paper
