### 2023-01-23 #112 @matbesancon
that's why the logs was breaking early
