+++ paper/header.tex
@@ -1,18 +1,17 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
-\title{GraphPPL.jl: A Julia Package for probabilistic model and inference constraints specification}
+\title{GraphPPL.jl: A Julia package for probabilistic model and Bethe Free Energy optimization constraints specification}
 
 \author[1]{Dmitry Bagaev}
 \author[1]{Bert de Vries}
-\affil[1]{Technical University of Eindhoven}
+\affil[1]{Technical University of Eindhoven, BIASlab}
 
-\keywords{probabilistic programming, variational Bayesian inference, factor graphs, Free Energy optimization}
+\keywords{probabilistic programming, Bayesian inference, variational inference, factor graphs, bethe free energy, graphical models, optimization}
 
 \hypersetup{
-pdftitle = {GraphPPL.jl: A Julia Package for probabilistic model and inference constraints specification},
-pdfsubject = {JuliaCon 2022 Proceedings},
+pdftitle = {GraphPPL.jl: A Julia package for probabilistic model and Bethe Free Energy optimization constraints specification},
+pdfsubject = {JuliaCon 2019 Proceedings},
 pdfauthor = {Dmitry Bagaev, Bert de Vries},
-pdfkeywords = {probabilistic programming, variational Bayesian inference, factor graphs, Free Energy optimization},
+pdfkeywords = {probabilistic programming, Bayesian inference, variational inference, factor graphs, bethe free energy, graphical models, optimization},
 }
-
 
