### 2023-01-25 #123 @whedon
[REVIEW]: GraphPPL.jl: A Julia Package for probabilistic model and inference constraints specification
**Submitting author:** @bvdmitri (<a href="http://orcid.org/0000-0001-9655-7986">Dmitry Bagaev</a>)
**Repository:** <a href="https://github.com/biaslab/GraphPPL.jl" target ="_blank">https://github.com/biaslab/GraphPPL.jl</a>
**Version:**

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [ ] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
