### 2023-02-09 #123 @theogf
I have started reviewing the abstract and the package and before I go further, there are some major issues for acceptance:
- In the docs, the examples do not work. There is no `@model` in `GraphPPL.jl`.
- If it is somewhere else it should be mentioned in the examples (and the docs statements should be corrected)
- There is close to zero test on the main functionality of the package

See generally the missing points in Functionality and Documentation in my checklist.

I have additional comments regarding the text but they can wait.
