### 2023-02-27 #123 @bvdmitri
Hey @theogf , your concerns are valid. That happened because we splitted up (it was quite some time since the initial submission) our repository and now the `@model` macro lives in the `RxInfer` repository. So the repository is currently not in a very good shape. The `GraphPPL` is being refactored by a PhD student of ours to allow more features, tests and be self isolated from the `RxInfer`. This work, however, will take some time :/
