### 2023-02-27 #123 @matbesancon
@bvdmitri thanks for the response. Given the context, could it be worth submitting the proceeding later, when the package is more mature and stabilized in its new version?
