### 2023-03-10 3d71dfe
Initial implementation of composite graphs
