+++ Project.toml
@@ -4,7 +4,14 @@ authors = ["Dmitry Bagaev <bvdmitri@gmail.com>"]
 version = "3.1.0"
 
 [deps]
+Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
+Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
+Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
+MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
+SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 TupleTools = "9d95972d-f1c8-5527-a6e0-b4b365fa01f6"
 
 [compat]
