### 2023-03-10 bcaefc4
Initial implementation of composite graphs
