+++ src/context_graphs.jl
@@ -1,0 +1,247 @@
+using Graphs
+using MetaGraphsNext
+import Base: put!, haskey, gensym, getindex
+import MacroTools: postwalk, capture
+using GraphPlot, Compose
+import Cairo
+
+struct Context
+    prefix :: String
+    contents :: Dict{Symbol, Union{Symbol, Context}}
+end
+
+name(f::Function) = String(Symbol(f))
+vals(kwargs::Base.Pairs) = [entry for entry in values(kwargs)]
+
+Context(prefix::String) = Context(prefix, Dict())    
+Context(parent::Context, model_name::String) = Context(parent.prefix * model_name * "_")
+Context(parent::Context, model_name::Function) = Context(parent, name(model_name))
+Context() = Context("")
+
+haskey(context :: Context, key::Symbol) = haskey(context.contents, key)
+
+gensym(context :: Context, key::Symbol) = gensym(Symbol(context.prefix * String(key)))
+gensym(context :: Context, key::Function) = gensym(Symbol(context.prefix * name(key)))
+
+getindex(c::Context, key::Symbol) = c.contents[key]
+
+
+function pprint(str:: String, indent :: Int)
+    print("   "^indent * str)
+end
+
+function pprint(context :: Context, indent=0)
+    println("Context " * context.prefix * "{")
+    for pair in context.contents
+        pprint(pair, indent+1)
+    end
+    pprint("} \n", indent)
+end
+
+function pprint(pair :: Pair{Symbol, Union{Symbol, Context}}, indent=0)
+    pprint(String(pair[1]) * " : ", indent)
+    pprint(pair[2], indent)
+end
+
+function pprint(symbol :: Symbol, indent=0)
+   println(String(symbol))
+end
+
+abstract type NodeType end
+
+struct Composite <: NodeType end
+struct Atomic <: NodeType end
+
+
+NodeType(::Type) = Atomic()
+NodeType(::Function) = Atomic()
+
+function Base.put!(context :: Context, name :: Symbol, variable :: Union{Symbol, Context})
+    if haskey(context, name)
+        throw(ErrorException("Variable " * String(name) * " in Context " * context.prefix * " is duplicate."))
+    end
+    context.contents[name] = variable
+end
+
+
+function create_model(interfaces...)
+    model = MetaGraph(
+        Graph(),
+        Label = Symbol,
+        VertexData = Tuple{Bool, String},
+        graph_data = Context(),
+        EdgeData = Symbol
+    )
+    for interface in interfaces
+        model[interface] = (true, String(interface))
+        put!(model[], interface, interface)
+
+    end
+    return model
+end
+
+function add_variable_node!(model:: MetaGraph, context :: Context, variable_id::Symbol)
+    variable_symbol = gensym(context, variable_id)
+    put!(context, variable_id, variable_symbol)
+    model[variable_symbol] = (true, String(variable_id))
+    return variable_symbol
+end
+
+function add_atomic_factor_node!(model :: MetaGraph, context :: Context, node_name::Union{Function, Symbol})
+    node_id = gensym(context, node_name)
+    model[node_id] = (false, name(node_name))
+    put!(context, node_id, node_id)
+    return node_id
+end
+
+function add_composite_factor_node!(parent_context :: Context, context :: Context, node_name::Union{Function, Symbol})
+    node_id = gensym(parent_context, node_name)
+    put!(parent_context, node_id, context)
+    return node_id
+end
+
+getorcreate(model::MetaGraph, context::Context, edge::Symbol) = get(() -> add_variable_node!(model, context, edge), context, edge)
+    # if !(haskey(context, edge))
+    #     return add_variable_node!(model, context, name)
+    # end
+    # return context[edge]
+
[…]
