### 2023-03-22 a77bbdc
Graph engine implemented and tested
