+++ test/graph_engine.jl
@@ -1,0 +1,475 @@
+module test_modular_graphs
+
+using Test
+using GraphPPL
+using Graphs
+using MetaGraphsNext
+using TestSetExtensions
+
+@testset "graph_engine" begin
+
+    @testset "model constructor" begin
+        import GraphPPL: Model, NodeLabel, NodeData, EdgeData, Context
+
+        g = MetaGraph(
+            Graph(),
+            Label = NodeLabel,
+            VertexData = NodeData,
+            graph_data = Context(),
+            EdgeData = Symbol,
+        )
+        @test typeof(Model(g)) == Model
+
+        @test_throws MethodError Model()
+    end
+
+    @testset "setindex!(::Model, ::NodeData, ::NodeLabel)" begin
+        import GraphPPL: create_model, NodeLabel, NodeData
+
+        model = create_model()
+        model[NodeLabel(:μ, 1)] = NodeData(true, :μ)
+        @test GraphPPL.nv(model) == 1 && GraphPPL.ne(model) == 0
+
+        @test_throws MethodError model[0] = 1
+
+        @test_throws MethodError model["string"] = NodeData(false, "string")
+        model[NodeLabel(:x, 2)] = NodeData(true, :x)
+        @test GraphPPL.nv(model) == 2 && GraphPPL.ne(model) == 0
+    end
+
+    @testset "setindex!(::Model, ::EdgeData, ::NodeLabel, ::NodeLabel)" begin
+        import GraphPPL: create_model, NodeLabel, NodeData, EdgeLabel
+
+        model = create_model()
+        μ = NodeLabel(:μ, 1)
+        x = NodeLabel(:x, 2)
+        model[μ] = NodeData(true, :μ)
+        model[x] = NodeData(true, :x)
+        model[μ, x] = EdgeLabel(:interface)
+        @test GraphPPL.ne(model) == 1
+
+        @test_throws MethodError model[0, 1] = 1
+
+        @test_throws KeyError model[μ, NodeLabel(:x, 100)] = EdgeLabel(:if)
+    end
+
+    @testset "getindex(::Model, ::NodeLabel)" begin
+        import GraphPPL: create_model, NodeLabel, NodeData
+
+        model = create_model(:x)
+        @test model[NodeLabel(:x, 1)] == NodeData(true, :x)
+        @test_throws KeyError model[NodeLabel(:x, 10)]
+        @test_throws MethodError model[0]
+
+        @test_throws KeyError model[NodeLabel(:x, 2)]
+    end
+
+    @testset "increase_count(::Model)" begin
+        import GraphPPL: create_model, increase_count
+        model = create_model()
+
+        increase_count(model)
+
+        @test model.counter == 1
+        increase_count(model)
+        @test model.counter == 2
+    end
+
+    @testset "nv_ne(::Model)" begin
+        import GraphPPL: create_model, nv, ne, NodeData, NodeLabel, EdgeLabel
+
+        model = create_model()
+        @test nv(model) == 0
+        @test ne(model) == 0
+
+        model[NodeLabel(:a, 1)] = NodeData(true, :a)
+        model[NodeLabel(:b, 2)] = NodeData(false, :b)
+        @test nv(model) == 2
+        @test ne(model) == 0
+
+        model[NodeLabel(:a, 1), NodeLabel(:b, 2)] = EdgeLabel(:edge)
+        @test nv(model) == 2
+        @test ne(model) == 1
+    end
+
+    @testset "gensym(::Model, ::Symbol)" begin
+        import GraphPPL: create_model, gensym, NodeLabel
+
+        model = create_model()
+        first_sym = gensym(model, :x)
+        @test typeof(first_sym) == NodeLabel
+
[…]
