+++ src/model.jl
@@ -1,8 +1,15 @@
 
 import MacroTools: @capture, postwalk, prewalk, walk
 
-function conditioned_walk(f, condition_skip, condition_apply, x) 
-    walk(x, x -> condition_skip(x) ? x : condition_apply(x) ? f(x) : conditioned_walk(f, condition_skip, condition_apply, x), identity)
+function conditioned_walk(f, condition_skip, condition_apply, x)
+    walk(
+        x,
+        x ->
+            condition_skip(x) ? x :
+            condition_apply(x) ? f(x) :
+            conditioned_walk(f, condition_skip, condition_apply, x),
+        identity,
+    )
 end
 
 """
@@ -11,10 +18,10 @@ end
 This function forces `Expr` or `Symbol` to be quoted.
 """
 fquote(expr::Symbol) = Expr(:quote, expr)
-fquote(expr::Int)    = expr
-fquote(expr::Expr)   = expr
+fquote(expr::Int) = expr
+fquote(expr::Expr) = expr
 
-is_kwargs_expression(x)       = false
+is_kwargs_expression(x) = false
 is_kwargs_expression(x::Expr) = x.head === :parameters
 
 """
@@ -26,9 +33,9 @@ This function parses variable id and returns a tuple of 3 different representati
 3. Full variable identificator (used in model as a variable id)
 """
 function parse_varexpr(varexpr::Symbol)
-    varexpr  = varexpr
+    varexpr = varexpr
     short_id = varexpr
-    full_id  = varexpr
+    full_id = varexpr
     return varexpr, short_id, full_id
 end
 
@@ -36,15 +43,23 @@ function parse_varexpr(varexpr::Expr)
 
     # TODO: It might be handy to have this feature in the future for e.g. interacting with UnPack.jl package
     # TODO: For now however we fallback to a more informative error message since it is not obvious how to parse such expressions yet
-    @capture(varexpr, (tupled_ids__, )) && 
-        error("Multiple variable declarations, definitions and assigments are forbidden within @model macro. Try to split $(varexpr) into several independent statements.")
+    @capture(varexpr, (tupled_ids__,)) && error(
+        "Multiple variable declarations, definitions and assigments are forbidden within @model macro. Try to split $(varexpr) into several independent statements.",
+    )
 
-    @capture(varexpr, id_[idx__]) || 
-        error("Variable identificator can be in form of a single symbol (x ~ ...) or indexing expression (x[i] ~ ...)")
-   
-    varexpr  = varexpr
+    @capture(varexpr, id_[idx__]) || error(
+        "Variable identificator can be in form of a single symbol (x ~ ...) or indexing expression (x[i] ~ ...)",
+    )
+
+    varexpr = varexpr
     short_id = id
-    full_id  = Expr(:call, :Symbol, fquote(id), Expr(:quote, :_), Expr(:quote, Symbol(join(idx, :_))))
+    full_id = Expr(
+        :call,
+        :Symbol,
+        fquote(id),
+        Expr(:quote, :_),
+        Expr(:quote, Symbol(join(idx, :_))),
+    )
 
     return varexpr, short_id, full_id
 end
@@ -69,42 +84,53 @@ end
 
 function __normalize_arg(backend, model, arg)
     if @capture(arg, constvar(arguments__))
-        return write_constvar_expression(backend, model, gensym(:anonymous_constvar), arguments)
+        return write_constvar_expression(
+            backend,
+            model,
+            gensym(:anonymous_constvar),
+            arguments,
+        )
     elseif @capture(arg, constvar.(arguments__))
[…]
