+++ src/GraphPPL.jl
@@ -6,6 +6,7 @@ include("utils.jl")
 include("model.jl")
 include("constraints.jl")
 include("meta.jl")
-include("context_graphs.jl")
+include("graph_engine.jl")
+include("model_macro.jl")
 
 end # module
