+++ Project.toml
@@ -12,6 +12,7 @@ Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
+TestSetExtensions = "98d24dd4-01ad-11ea-1b02-c9a08f80db04"
 TupleTools = "9d95972d-f1c8-5527-a6e0-b4b365fa01f6"
 
 [compat]

[…]
