### 2023-03-22 d910108
Graph engine implemented and tested. Alpha model macro implemented. Formatter added
Implemented Graph Engine and alpha version of the Model macro. This commit is huge, as it is the initial commit to the 4.0.0 version of GraphPPL. From hereon forward we'll commit regularly. Formatter is added. `src/graph_engine.jl~ is rigorously tested. Tests for `src/model_macro.jl` will be added in future commits.
