+++ Makefile
@@ -1,0 +1,8 @@
+scripts_init:
+	julia --startup-file=no --project=scripts/ -e 'using Pkg; Pkg.instantiate(); Pkg.update(); Pkg.precompile();'
+
+lint: scripts_init ## Code formating check
+	julia --startup-file=no --project=scripts/ scripts/format.jl
+
+format: scripts_init ## Code formating run
+	julia --startup-file=no --project=scripts/ scripts/format.jl --overwrite

[…]
