+++ .github/workflows/ci.yml
@@ -1,7 +1,11 @@
 name: CI
 on:
   pull_request:
+    types: [ready_for_review,reopened,synchronize]
   push:
+    branches:
+        - 'master'
+      tags: '*'
   schedule:
     - cron: '44 9 16 * *' # run the cron job one time per month
 jobs:
