### 2023-03-23 952093a
Update CI routine to only run on pull requests and pushes to master
