+++ .github/workflows/ci.yml
@@ -2,6 +2,8 @@ name: CI
 on:
   pull_request:
     types: [ready_for_review,reopened,synchronize]
+  pull_request_review:
+    types: [submitted,edited]
   push:
     branches:
         - 'master'
