+++ .github/workflows/ci.yml
@@ -6,8 +6,8 @@ on:
     types: [submitted,edited]
   push:
     branches:
-        - 'master'
-      tags: '*'
+      - 'master'
+    tags: '*'
   schedule:
     - cron: '44 9 16 * *' # run the cron job one time per month
 jobs:
