+++ src/graph_engine.jl
@@ -4,7 +4,10 @@ import Base: put!, haskey, gensym, getindex, getproperty, setproperty!, setindex
 using GraphPlot, Compose
 import Cairo
 
-
+"""
+The Model struct contains all information about the Factor Graph and contains a MetaGraph object and a counter. 
+The counter is implemented because it allows for an efficient `gensym`` implementation
+"""
 struct Model
     graph::MetaGraph
     counter::Base.RefValue{Int64}
@@ -24,6 +27,17 @@ struct EdgeLabel
     name::Symbol
 end
 
+"""
+    Model(graph::MetaGraph)
+
+A structure representing a probabilistic graphical model. It contains a `MetaGraph` object
+representing the factor graph and a `Base.RefValue{Int64}` object to keep track of the number
+of nodes in the graph.
+
+Fields:
+- `graph`: A `MetaGraph` object representing the factor graph.
+- `counter`: A `Base.RefValue{Int64}` object keeping track of the number of nodes in the graph.
+"""
 Model(graph::MetaGraph) = Model(graph, Base.RefValue(0))
 
 
@@ -56,6 +70,20 @@ increase_count(model::Model) = Base.setproperty!(model, :counter, model.counter 
 Graphs.nv(model::Model) = Graphs.nv(model.graph)
 Graphs.ne(model::Model) = Graphs.ne(model.graph)
 
+
+"""
+    gensym(model::Model, name::Symbol)
+
+Generate a new `NodeLabel` object with a unique identifier based on the specified name and the
+number of nodes already in the model.
+
+Arguments:
+- `model`: A `Model` object representing the probabilistic graphical model.
+- `name`: A symbol representing the name of the node.
+
+Returns:
+A new `NodeLabel` object with a unique identifier.
+"""
 function gensym(model::Model, name::Symbol)
     increase_count(model)
     return NodeLabel(name, model.counter)
@@ -132,6 +160,20 @@ end
 Base.put!(context::Context, name::NodeLabel, variable::Union{NodeLabel,Context}) =
     Base.put!(context, to_symbol(name), variable)
 
+
+
+"""
+create_model(interfaces...)
+
+Create a new probabilistic graphical model with the specified interfaces. Each interface is
+represented as a variable node in the factor graph.
+
+Arguments:
+- `interfaces`: A variable number of symbols representing the names of the interface variables.
+
+Returns:
+A `Model` object representing the probabilistic graphical model.
+"""
 function create_model(interfaces...)
     model = MetaGraph(
         Graph(),
@@ -158,11 +200,40 @@ end
 
 getorcreate!(model::Model, context::Context, edge) = edge
 getorcreate!(model::Model, something) = getorcreate!(model, context(model), something)
+"""
+    Creates a variable node for the given `edge` in the `model` under the given `context`. 
+    If a variable node for the `edge` already exists, returns it. Otherwise, creates a new 
+    variable node and adds it to the `model` and the `context`. Returns the variable node.
+"""
 getorcreate!(model::Model, context::Context, edge::Symbol) =
     get!(() -> add_variable_node!(model, context, edge), context.contents, edge)
+
+"""
+    Creates variable nodes for each element in the `edges` collection in the `model` under the given `context`. 
+    If a variable node for an element already exists, returns it. Otherwise, creates a new variable node for 
+    that element and adds it to the `model` and the `context`. Returns a collection of variable nodes,
+    either a tuple or a vector, depending on the type of the input `edges`.
+"""
 getorcreate!(model::Model, context::Context, edges::Union{Tuple,AbstractArray}) =
     map((edge) -> getorcreate!(model, context, edge), edges)
 
+
+"""
+Add a variable node to the model with the given ID.
+
[…]
