+++ test/model_macro.jl
@@ -14,6 +14,16 @@ using MacroTools
 
         @test missing_interfaces(sum, Val(3), (in1 = :x, in2 = :y)) == [:out]
         @test missing_interfaces(sum, Val(3), (out = :y,)) == [:in1, :in2]
+        @test missing_interfaces(sum, Val(3), Dict()) == [:in1, :in2, :out]
+
+        GraphPPL.interfaces(::typeof(prod), ::Val{0}) = []
+        @test missing_interfaces(prod, Val(0), (in1 = :x, in2 = :y)) == []
+
+        GraphPPL.interfaces(::typeof(sin), ::Val{2}) = (:a, :b)
+        @test missing_interfaces(sin, Val(2), (a = 1, b = 2)) == []
+
+        GraphPPL.interfaces(::typeof(div), ::Val{2}) = (:in1, :in2, :out)
+        @test missing_interfaces(div, Val(2), (in1 = 1, in2 = 2, out = 3, test = 4)) == []
     end
 
     @testset "generate_get_or_create_expression(::AbstractArray)" begin
@@ -29,6 +39,19 @@ using MacroTools
         @test_throws MethodError GraphPPL.generate_get_or_create_expression(input)
     end
 
+    @testset "generate_get_or_create_expression(::Expr)" begin
+        import GraphPPL: generate_get_or_create_expression
+        input = Expr(:tuple, [:x, :y])
+        output = quote
+            x = (@isdefined(x)) ? x : GraphPPL.getorcreate!(model, context, :x)
+            y = (@isdefined(y)) ? y : GraphPPL.getorcreate!(model, context, :y)
+        end
+        @test_expression_generating(generate_get_or_create_expression(input), output)
+
+        input = Expr(:tuple, [:x, 1])
+        @test_throws MethodError GraphPPL.generate_get_or_create_expression(input)
+    end
+
     @testset "generate_get_or_create_expression(::Symbol)" begin
         input = :x
         output = quote
@@ -38,6 +61,33 @@ using MacroTools
 
         input = 1
         @test_throws MethodError GraphPPL.generate_get_or_create_expression(input)
+    end
+
+    @testset "is_kwargs_expression(::AbstractArray)" begin
+        import GraphPPL: is_kwargs_expression
+
+        func_def = :(foo(a, b))
+        @capture(func_def, (f_(args__)))
+        @test !is_kwargs_expression(args)
+
+        func_def = :(foo(a = a, b = b))
+        @capture(func_def, (f_(args__)))
+        @test is_kwargs_expression(args)
+
+        empty_args = Expr(:tuple)
+        @test !is_kwargs_expression(empty_args)
+
+        empty_args = :(foo())
+        @capture(empty_args, (f_(args__)))
+        @test !is_kwargs_expression(args)
+
+        mixed_args = :(foo(a, b, c = c))
+        @capture(mixed_args, (f_(args__)))
+        @test !is_kwargs_expression(args)
+
+        nested_args = :(foo(a = a, b = b, c = (d = d)))
+        @capture(nested_args, (f_(args__)))
+        @test is_kwargs_expression(args)
     end
 
     @testset "keyword_expressions_to_named_tuple" begin
@@ -50,6 +100,7 @@ using MacroTools
     @testset "generate_make_node_call(::Any, NamedTuple)" begin
         import GraphPPL: create_model, generate_make_node_call
 
+        #test default functionality
         input = (sum, (in = :x, out = :y))
         output = quote
             x = (@isdefined(x)) ? x : GraphPPL.getorcreate!(model, context, :x)
@@ -60,6 +111,7 @@ using MacroTools
 
         @test_expression_generating(generate_make_node_call(input...), output)
 
+        #test expression functionality
         input = (sum, (in = :([x]), out = :y))
         output = quote
[…]
