+++ src/GraphPPL.jl
@@ -6,7 +6,9 @@ include("utils.jl")
 include("model.jl")
 include("constraints.jl")
 include("meta.jl")
+include("resizable_array.jl")
 include("graph_engine.jl")
 include("model_macro.jl")
+
 
 end # module
