+++ src/graph_engine.jl
@@ -1,12 +1,22 @@
 using Graphs
 using MetaGraphsNext
-import Base: put!, haskey, gensym, getindex, getproperty, setproperty!, setindex!
+import Base:
+    put!,
+    haskey,
+    gensym,
+    getindex,
+    getproperty,
+    setproperty!,
+    setindex!,
+    length,
+    size,
+    resize!
 using GraphPlot, Compose
 import Cairo
 
 """
 The Model struct contains all information about the Factor Graph and contains a MetaGraph object and a counter. 
-The counter is implemented because it allows for an efficient `gensym`` implementation
+The counter is implemented because it allows for an efficient `gensym` implementation
 """
 struct Model
     graph::MetaGraph
@@ -16,7 +26,16 @@ end
 struct NodeLabel
     name::Symbol
     index::Int64
+    variable_type::UInt8
+    variable_index::Union{Int64,NTuple{N,Int64} where N}
 end
+
+NodeLabel(
+    name::Symbol,
+    index::Int64,
+    variable_type::Int64,
+    variable_index::Union{Int64,NTuple{N,Int64} where N},
+) = NodeLabel(name, index, UInt8(variable_type), variable_index)
 
 struct NodeData
     is_variable::Bool
@@ -80,58 +99,92 @@ number of nodes already in the model.
 Arguments:
 - `model`: A `Model` object representing the probabilistic graphical model.
 - `name`: A symbol representing the name of the node.
+- `variable_type`: A UInt8 representing the type of the variable. 0 = factor, 1 = individual variable, 2 = vector variable, 3 = tensor variable
+- `index`: An integer or tuple of integers representing the index of the variable.
 
 Returns:
 A new `NodeLabel` object with a unique identifier.
 """
-function gensym(model::Model, name::Symbol)
+function gensym(
+    model::Model,
+    name::Symbol,
+    variable_type::UInt8 = UInt8(0),
+    index::Union{Int64,NTuple{N,Int64} where N} = 0,
+)
     increase_count(model)
-    return NodeLabel(name, model.counter)
+    return NodeLabel(name, model.counter, variable_type, index)
 end
 
-gensym(model::Model, name) = gensym(model::Model, Symbol(name))
+gensym(model::Model, name::Symbol, index::Nothing) = gensym(model, name, UInt8(1), 0)
+gensym(model::Model, name::Symbol, index::Int64) = gensym(model, name, UInt8(2), index)
+gensym(model::Model, name::Symbol, index::NTuple{N,Int64} where {N}) =
+    gensym(model, name, UInt8(3), index)
+gensym(model::Model, name::Symbol, index::Tuple) = throw(ArgumentError("Index, if provided, must be an integer or tuple of integers, not a $(typeof(index))"))
+
+gensym(model::Model, name, index = nothing) = gensym(model::Model, Symbol(name), index)
 
 to_symbol(id::NodeLabel) = Symbol(String(id.name) * "_" * string(id.index))
 
+
 struct Context
     prefix::String
-    contents::Dict{Symbol,Union{NodeLabel,Context}}
+    individual_variables::Dict{Symbol,NodeLabel}
+    vector_variables::Dict{Symbol,ResizableArray{NodeLabel}}
+    tensor_variables::Dict{Symbol,ResizableArray}
+    factor_nodes::Dict{NodeLabel,Union{NodeLabel,Context}}
 end
 
 
 name(f::Function) = String(Symbol(f))
 
-Context(prefix::String) = Context(prefix, Dict())
+Context(prefix::String) = Context(prefix, Dict(), Dict(), Dict(), Dict())
 Context(parent::Context, model_name::String) = Context(parent.prefix * model_name * "_")
[…]

[…]
