+++ test/graph_engine.jl
@@ -7,9 +7,8 @@ using MetaGraphsNext
 using TestSetExtensions
 
 @testset "graph_engine" begin
-
     @testset "model constructor" begin
-        import GraphPPL: Model, NodeLabel, NodeData, EdgeData, Context
+        import GraphPPL: Model, NodeLabel, NodeData, Context
 
         g = MetaGraph(
             Graph(),
@@ -27,22 +26,22 @@ using TestSetExtensions
         import GraphPPL: create_model, NodeLabel, NodeData
 
         model = create_model()
-        model[NodeLabel(:μ, 1)] = NodeData(true, :μ)
+        model[NodeLabel(:μ, 1, 1, 0)] = NodeData(true, :μ)
         @test GraphPPL.nv(model) == 1 && GraphPPL.ne(model) == 0
 
         @test_throws MethodError model[0] = 1
 
         @test_throws MethodError model["string"] = NodeData(false, "string")
-        model[NodeLabel(:x, 2)] = NodeData(true, :x)
+        model[NodeLabel(:x, 2, 1, 0)] = NodeData(true, :x)
         @test GraphPPL.nv(model) == 2 && GraphPPL.ne(model) == 0
     end
 
-    @testset "setindex!(::Model, ::EdgeData, ::NodeLabel, ::NodeLabel)" begin
+    @testset "setindex!(::Model, ::EdgeLabel, ::NodeLabel, ::NodeLabel)" begin
         import GraphPPL: create_model, NodeLabel, NodeData, EdgeLabel
 
         model = create_model()
-        μ = NodeLabel(:μ, 1)
-        x = NodeLabel(:x, 2)
+        μ = NodeLabel(:μ, 1, 1, 0)
+        x = NodeLabel(:x, 2, 1, 0)
         model[μ] = NodeData(true, :μ)
         model[x] = NodeData(true, :x)
         model[μ, x] = EdgeLabel(:interface)
@@ -50,18 +49,18 @@ using TestSetExtensions
 
         @test_throws MethodError model[0, 1] = 1
 
-        @test_throws KeyError model[μ, NodeLabel(:x, 100)] = EdgeLabel(:if)
+        @test_throws KeyError model[μ, NodeLabel(:x, 1, 100, 0)] = EdgeLabel(:if)
     end
 
     @testset "getindex(::Model, ::NodeLabel)" begin
         import GraphPPL: create_model, NodeLabel, NodeData
 
-        model = create_model(:x)
-        @test model[NodeLabel(:x, 1)] == NodeData(true, :x)
-        @test_throws KeyError model[NodeLabel(:x, 10)]
+        model = create_model()
+        label = NodeLabel(:x, 1, 1, 0)
+        model[label] = NodeData(true, :x)
+        @test model[label] == NodeData(true, :x)
+        @test_throws KeyError model[NodeLabel(:x, 1, 10, 0)]
         @test_throws MethodError model[0]
-
-        @test_throws KeyError model[NodeLabel(:x, 2)]
     end
 
     @testset "increase_count(::Model)" begin
@@ -82,12 +81,12 @@ using TestSetExtensions
         @test nv(model) == 0
         @test ne(model) == 0
 
-        model[NodeLabel(:a, 1)] = NodeData(true, :a)
-        model[NodeLabel(:b, 2)] = NodeData(false, :b)
+        model[NodeLabel(:a, 1, 1, 0)] = NodeData(true, :a)
+        model[NodeLabel(:b, 2, 1, 0)] = NodeData(false, :b)
         @test nv(model) == 2
         @test ne(model) == 0
 
-        model[NodeLabel(:a, 1), NodeLabel(:b, 2)] = EdgeLabel(:edge)
[…]
