+++ test/graph_engine.jl
@@ -232,7 +232,7 @@ using TestSetExtensions
         @test child_context[:in1].name == :x
     end
 
-    
+
     @testset "getorcreate!" begin
         import GraphPPL: create_model, getorcreate!, Context
 
@@ -276,10 +276,10 @@ using TestSetExtensions
         @test nv(model) == 3
         @test context(model)[:mv][2, 4] == mv
 
-       
+
 
     end
-    
+
 
     @testset "add_variable_node!" begin
         import GraphPPL: create_model, add_variable_node!, context
@@ -311,9 +311,9 @@ using TestSetExtensions
         #add a second vector variable to the model
         node_id = add_variable_node!(model, ctx, :x, 1)
         @test nv(model) == 2 &&
-                haskey(ctx, :x) &&
-                ctx[:x][1] == node_id &&
-                length(ctx[:x]) == 2
+              haskey(ctx, :x) &&
+              ctx[:x][1] == node_id &&
+              length(ctx[:x]) == 2
 
         #add a tensor variable to the model
         model = create_model()
@@ -357,7 +357,7 @@ using TestSetExtensions
 
     @testset "add_composite_factor_node!" begin
         import GraphPPL: create_model, add_composite_factor_node!, context
-    
+
         # Add a composite factor node to the model
         model = create_model()
         parent_ctx = context(model)
@@ -365,20 +365,26 @@ using TestSetExtensions
         add_variable_node!(model, child_ctx, :x)
         add_variable_node!(model, child_ctx, :y)
         node_id = add_composite_factor_node!(model, parent_ctx, child_ctx, :f)
-        @test nv(model) == 2 && haskey(parent_ctx.factor_nodes, node_id) &&
-              parent_ctx.factor_nodes[node_id] === child_ctx && length(child_ctx.individual_variables) == 2
+        @test nv(model) == 2 &&
+              haskey(parent_ctx.factor_nodes, node_id) &&
+              parent_ctx.factor_nodes[node_id] === child_ctx &&
+              length(child_ctx.individual_variables) == 2
 
-    
+
         # Add a composite factor node with a different name
         node_id = add_composite_factor_node!(model, parent_ctx, child_ctx, :g)
-        @test nv(model) == 2 && haskey(parent_ctx.factor_nodes, node_id) &&
-              parent_ctx.factor_nodes[node_id] === child_ctx && length(child_ctx.individual_variables) == 2
-    
+        @test nv(model) == 2 &&
+              haskey(parent_ctx.factor_nodes, node_id) &&
+              parent_ctx.factor_nodes[node_id] === child_ctx &&
+              length(child_ctx.individual_variables) == 2
+
         # Add a composite factor node with an empty child context
         empty_ctx = Context()
         node_id = add_composite_factor_node!(model, parent_ctx, empty_ctx, :h)
-        @test nv(model) == 2 && haskey(parent_ctx.factor_nodes, node_id) &&
-              parent_ctx.factor_nodes[node_id] === empty_ctx && length(empty_ctx.individual_variables) == 0
+        @test nv(model) == 2 &&
+              haskey(parent_ctx.factor_nodes, node_id) &&
+              parent_ctx.factor_nodes[node_id] === empty_ctx &&
+              length(empty_ctx.individual_variables) == 0
     end
 
     @testset "add_edge!(::Model, ::NodeLabel, ::NodeLabel, ::Symbol)" begin
